0.5.6.5, fix #444, fix #446

+++ src/main/java/pl/asie/charset/lib/utils/redstone/RedstoneUtils.java
@@ -60,7 +60,7 @@ public final class RedstoneUtils {
 		for (int i = 0; i < handler.getSlots(); i++) {
 			ItemStack stack = handler.getStackInSlot(i);
 			if (!stack.isEmpty()) {
-				value += stack.getCount() / Math.min(stack.getMaxStackSize(), handler.getSlotLimit(i));
+				value += stack.getCount() / (float) Math.min(stack.getMaxStackSize(), handler.getSlotLimit(i));
 				count++;
 			}
 		}
